Тест на GeometryWriter, рефакторинг скрипта карты
