+++ ApplicationServices/Service/SpatialRelationsService.cs
@@ -6,12 +6,14 @@ public class SpatialRelationsService
 {
     public SpatialRelationsInfo GetSpatialRelationsInfo(IGeometryPrimitive first, IGeometryPrimitive second)
     {
-        var relationsInfo = new SpatialRelationsInfo();
         var isIntersect = GeometryModels.Extensions.IntersectorExtension.Intersects(first, second);
         var isInside = GeometryModels.Extensions.InsiderExtension.IsInside(first, second);
 
-        relationsInfo.Intersecting = isIntersect;
-        relationsInfo.Inside = isInside;
+        var relationsInfo = new SpatialRelationsInfo
+        {
+            Intersecting = isIntersect,
+            Inside = isInside
+        };
         return relationsInfo;
     }
 }
