+++ src/components/ui/dropdownMenu/DropdownMenu.jsx
@@ -2,8 +2,8 @@ import React, {useState} from "react";
 import "./DropdownMenu.css"
 
 import {removeAllObjects} from '../../workArea/Map.js';
-import {newPolyline} from '../../workArea/Map.js';
-import {newPolygon} from '../../workArea/Map.js';
+import {polylineActions} from '../../workArea/MapObjectFactory.js';
+import {polygonActions} from '../../workArea/MapObjectFactory.js';
 
 function DropdownMenu () {
     const [isOpen, setIsOpen] = useState(false);  
@@ -16,9 +16,9 @@ function DropdownMenu () {
 
             <div className={'dropdown ' + (isOpen ? 'open' : 'closed')}>
                 <ul>
-                    <li><button className="DropBtn" onClick = {newPolyline}>Полилиния</button>
+                    <li><button className="DropBtn" onClick = {polylineActions}>Полилиния</button>
                     <span class="material-icons">timeline</span></li>
-                    <li><button className="DropBtn" onClick = {newPolygon}>Полигон</button>
+                    <li><button className="DropBtn" onClick = {polygonActions}>Полигон</button>
                     <span class="material-icons">polymer</span></li>
                     <li><button className="DropBtn" onClick = {removeAllObjects}>Очистка</button>
                     <span class="material-symbols-outlined">delete</span></li>
