+++ src/pages/mainPage/MainPage.jsx
@@ -1,7 +1,7 @@
 import React, { createRef, useState, } from "react";
 import "./MainPage.css"
 import Header from "../../components/header/Header.jsx"
-import Map1 from "../../components/workArea/Map.js"
+import GemPixMap from "../../components/workArea/Map.js"
 import Footer from "../../components/footer/Footer.jsx"
 import Info from "../../components/ui/Info/Info"
 import {clearTempObjects} from '../../components/workArea/Map.js';
@@ -10,7 +10,7 @@ function MainPage() {
     return (
     <body onClick = {clearTempObjects}>
      <Header/>
-     <Map1/>
+     <GemPixMap/>
     
      <Footer/>
     </body>
